refactor: 별점 관련 로직 이벤트 핸들러 분리 및 함수 분리
